Rename "atomic" to "consistent".
